requirements.txt
python-docx
tkinter
